+++ NEWS.md
@@ -1,5 +1,8 @@
 # httr2 (development version)
 
+* `req_retry()` now defaults to `max_tries = 2` with a message.
+  Set to `max_tries = 1` to disable retries.
+
 * Errors thrown during the parsing of an OAuth response now have a dedicated
   `httr2_oauth_parse` error class that includes the original response object
   (@atheriel, #596).